Acl 수정 아직 버그 있음. 그룹 버그 존재함. 동작만 확인

+++ teamKMS/src/main/java/com/devworker/kms/service/GroupServiceImpl.java
@@ -1,9 +1,11 @@
 package com.devworker.kms.service;
 
 import com.devworker.kms.dao.GroupDao;
+import com.devworker.kms.dic.PermissionType;
 import com.devworker.kms.dto.GroupDto;
 import com.devworker.kms.exception.NotExistException;
 import com.devworker.kms.repo.GroupRepo;
+import com.devworker.kms.util.AclUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
@@ -25,12 +27,14 @@ public class GroupServiceImpl implements GroupService{
 	
 	@Override
 	public int addGroup(GroupDto dto) {
+		AclUtil.checkPermission(PermissionType.CREATEGROUP);
 		GroupDao save = repo.save(dto.getDao());
 		return save.getId();
 	}
 	
 	@Override
 	public void deleteGroup(int id) {
+		AclUtil.checkPermission(PermissionType.DELETEGROUP);
 		repo.deleteById(id);
 	}
